+++ CHANGELOG.md
@@ -2,12 +2,16 @@
 
 ## TODO
 - Finish integration tests for history.
-- Add secondary write adpater.
 - Add Elastic Search option.
 - Add Denormalized view feature (copy column data over)
 
+## [1.3.5] 2-17-2018
+- Fixed an issue with sub optimized queries.
+- Added optional disconnect method to adapters.
+
 ## [1.3.4] 2-13-2018
 - Fixed issue with `.extend("clone"...)`.
+- Now supports secondary adapters for live backups.
 
 ## [1.3.3] 2-11-2018
 - Removed unnecessary argument from database adapter implimentation.